fix: break the code into modules
