+++ js/util.js
@@ -1,0 +1,21 @@
+/**
+ * Функция по получению случайного числа из диапазона
+ * @param {number} a - нижняя граница
+ * @param {number} b - верхняя граница
+ * @returns {number} result - случайный индекс элемента из массива
+ */
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+/**
+ * Функция по поиску случайного элемента в переданном массиве
+ * @param {object} elements - массив данных
+ * @returns {string} element - элемент массива
+ */
+const elementFromArray = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, elementFromArray};
